+++ tools/nevermore-cli-helpers/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](https://github.com/Quenty/Nevermore/compare/@quenty/nevermore-cli-helpers@1.15.0...@quenty/nevermore-cli-helpers@1.16.0) (2026-05-26)
+
+**Note:** Version bump only for package @quenty/nevermore-cli-helpers
+
+# [1.15.0](https://github.com/Quenty/Nevermore/compare/@quenty/nevermore-cli-helpers@1.14.0...@quenty/nevermore-cli-helpers@1.15.0) (2026-05-26)
+
+**Note:** Version bump only for package @quenty/nevermore-cli-helpers
+
+# [1.14.0](https://github.com/Quenty/Nevermore/compare/@quenty/nevermore-cli-helpers@1.13.0...@quenty/nevermore-cli-helpers@1.14.0) (2026-05-26)
+
+**Note:** Version bump only for package @quenty/nevermore-cli-helpers
+
 # [1.13.0](https://github.com/Quenty/Nevermore/compare/@quenty/nevermore-cli-helpers@1.12.0...@quenty/nevermore-cli-helpers@1.13.0) (2026-05-20)
 
 **Note:** Version bump only for package @quenty/nevermore-cli-helpers